Dont draw and move entities when game is paused
This is a quick solution because the entities will be drawn over gui. Better fix would be to draw the gui on top of the entities

+++ Game/Source/EnemyAir.cpp
@@ -38,15 +38,18 @@ bool EnemyAir::Start() {
 
 bool EnemyAir::Update()
 {
-	if (health <= 0) {
-		pbody->body->DestroyFixture(pbody->body->GetFixtureList());
-		app->entityManager->DestroyEntity(this);
-	}
+	if (app->scene->gamePaused == false) {
+		if (health <= 0) {
+			pbody->body->DestroyFixture(pbody->body->GetFixtureList());
+			app->entityManager->DestroyEntity(this);
+		}
 
-	FindPath();
-	Move();
-	RenderEntity();
-	UpdateAnim();
+		FindPath();
+		Move();
+		RenderEntity();
+		UpdateAnim();
+	}
+	
 	return true;
 }
